% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.availability.R
\name{data.availability}
\alias{data.availability}
\title{Checking which Datasets are available for Each Cancer}
\usage{
data.availability()
}
\value{
A matrix that contain all cancers and their available datasets. It is
available in the global enviroment (user's workspace). For convenience, an excel
file will also be generated in the working directory.
}
\description{
This function checks all the cancers that are registered in 'cbioportal.org' to
examine whether or not they contain RNA-seq, microRNA-seq, microarray(mRNA),
microarray(miRNA) and methylation datasets.
}
\details{
\tabular{lllll}{
Package: \tab cBioAutomatedTools \cr
Type: \tab Package \cr
Version: \tab 0.99.0 \cr
Date: \tab 2017-05-30 \cr
License: \tab Artistic-2.0 \cr
}

This function checks all the cancers that are registered in 'cbioportal.org' to
examine whether or not they contain RNA-seq, microRNA-seq, microarray(mRNA),
microarray(miRNA) and methylation datasets.
}
\author{
Arman Shahrisa, \email{shahrisa.arman@hotmail.com} [maintainer, copyright holder]

Maryam Tahmasebi Birgani, \email{tahmasebi-ma@ajums.ac.ir}
}
