% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expression.Data.For.a.Cancer.R
\name{Expression.Data.For.a.Cancer}
\alias{Expression.Data.For.a.Cancer}
\title{Checking Expression/methylation Profile for various subgroups of a cancer}
\usage{
Dataset.availability()
}
\value{
A matrix that contain all cancers and their available datasets. It is
available in the global enviroment (user's workspace). For convenience, an excel
file will also be generated in the working directory.
}
\description{
This function checks all the cancers from cbioportal.org to determine which
datasets they contain.
}
\details{
This function checks all the cancers that are registered in 'cbioportal.org' to
examine whether or not they contain RNA-seq, microRNA-seq, microarray(mRNA),
microarray(miRNA) and methylation datasets.
}
