% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.data.for.a.cancer.R
\name{process.data.for.a.cancer}
\alias{process.data.for.a.cancer}
\title{Checking Expression/methylation Profile for various subgroups of a cancer}
\usage{
process.data.for.a.cancer(genes, cancername, high.throughput.data.type,
data.presented.as = c("Frequency.Percentage", "Frequency.Ratio", "Mean.Expression", "Median"),
transposedHeatmap=FALSE, desired.case.list="None", genelimit="none",
resolution=600, RowCex=0.8, ColCex=0.8, heatmapMargines=c(10,10), cutoff="Default",
angle.for.heatmap.cancernames=45, heatmap.color = "RdBu", reverse.heatmap.color = TRUE,
resetOldExpressionProfile = TRUE, round=TRUE, top.genes = TRUE, validate.genes = TRUE,
Use.CancerCode.as.Name = FALSE, simplify.visulization=FALSE, simplifiction.cuttoff=FALSE)
}
\arguments{
\item{genes}{a list that contains at least one gene group}

\item{cancername}{a character string showing the desired cancer name. It containes standard cancer names
that can be found on cbioportal.org, such as \code{Acute Myeloid Leukemia (TCGA, NEJM 2013)}.}

\item{high.throughput.data.type}{a character string that is one of the following techniques: 'RNA-seq', 'microRNA-Seq',
'microarray.mRNA', 'microarray.microRNA' or 'methylation'.}

\item{data.presented.as}{a character vector that containes the statistical precedures users prefer the function to compute.
Default is \code{c("Frequency.Percentage", "Frequency.Ratio", "Mean.Expression", "Median")}. This will tell the function to
compute the following:
frequency precentage, which is the number of samples having value greather than specific cutoff divided by the total sample
size for each cancer;
frequency ratio, which shows the number of selected samples divided by the total number of samples that give the frequency
percentage
for each cancer-to know selecected and total sample sizes only;
Mean Expression, that contains mean value of selected samples for each cancer;
Median, which shows the median value of selected samples for each cancer.}

\item{transposedHeatmap}{a logical value that can exchange the rows and columns of heatmaps.}

\item{desired.case.list}{a numeric vector that contains the index of rdesired cancer subgroups. If set to be \code{FALSE},
function will ask the user to enter them during the process. The default value is \code{FALSE}.}

\item{genelimit}{If number of genes in at least one gene group is too much, this option can be use to limit the number of
genes to be shown on hitmap. For instance, \code{genelimit=50} will limit the heatmap to 50 genes showing the most variation.
The default value is \code{FALSE}.}

\item{resolution}{a number. This option can be used to adjust the resolution of the output heatmaps as 'dot per inch'. The
defalut value is 600.}

\item{RowCex}{a number that specifies letter size in heatmap row names.}

\item{ColCex}{a number that specifies letter size in heatmap column names.}

\item{heatmapMargines}{a numeric vectors that can be use to set heatmap margins. The default value is
\code{heatmapMargines=c(24,17)}.}

\item{cutoff}{a number that is used as limit to chose samples with values greather than this number. For methylation studies,
it is \code{observed/expected ration}, for the rest, it is \code{z-score}. The default value for methylation data is 0.6 while
gene expression studies use default value of 2.}

\item{angle.for.heatmap.cancernames}{a number that determines the angle with which the cancer names will be shown in heatmaps.
The default value is 45 degree.}

\item{heatmap.color}{a character string matches standard color names. The default value is "RdBu". "redgreen" is also very
popular in genomic studies. To see the rest, please type \code{display.brewer.all()}.}

\item{reverse.heatmap.color}{a logical value that reverses the color gradiant for heatmap.}

\item{resetOldExpressionProfile}{a logical value. This option can be used to modify heatmap, for instance change margin,
without obtaining data from internet again. If set to false, function will use tha values from global environment
(user's workspace) to draw heatmaps and save excel file. The default value is \code{TRUE}.}

\item{round}{a logical value that, if set to be \code{TRUE}, will force the function to rounds all the calculated values to
two decimal places.}

\item{top.genes}{a logical value that, if set as \code{TRUE}, cause the function to determine which gene containes the highest
value in each cancer subgroup.}

\item{validate.genes}{a logical value that, if set to be \code{TRUE}, enables pasrt of the function that checks whether or not
each gene has a record in different cancers. It also checks for alternative gene names that cbioportal might use instead of given
gene names.}

\item{Use.CancerCode.as.Name}{a logical value that tells the function to use abbreviated cancer names instead of complete
cancer names, if set to be \code{TRUE}. For example, \code{laml_tcga_pub} is the shortened name for
\code{Acute Myeloid Leukemia (TCGA, NEJM 2013)}.}

\item{simplify.visulization}{a logical value that tells the function whether or not to change values under
\code{simplifiction.cuttoff} to zero. It is designed for eye visulization purpose only. Therefore, it is not suited for
publications.}

\item{simplifiction.cuttoff}{a logical value that, if \code{simplify.visulization = TRUE}, needs to be set as a desired cuttoff
for \code{simplify.visulization}. It has the same unit as \code{cutoff}.}
}
\value{
a list that containes some or all of the following groups, based on what user chose: \code{Validation.Results},
\code{Frequency.Percentage}, \code{Top.Genes.of.Frequency.Percentage}, \code{Frequency.Ratio}, \code{Mean.Expression},
\code{Top.Genes.of.Mean.Expression}, \code{Median}, \code{Top.Genes.of.Median}. It also saves these groups in one excel
file for convenience. Based on preference, three heatmaps for \code{Frequency.Percentage}, \code{Mean.Expression} and
\code{Median} can be generated.
}
\description{
This function Obtaines the requested data for the given genes and subgroups of a cancer. It can check whether
or not all genes are included in subgroups of a cancer study and, if not, looks for the alternative gene names. Tha main
part of function calculates frequency percentage, frequency ratio, mean expression and median of sampleas greather than
specific value in selected subgroups of the cancer. Furthermore, it looks for the genes that have exhibited the highest
values in every cancer subgroup.
}
\details{
\tabular{lllll}{
Package: \tab cBioAutomatedTools \cr
Type: \tab Package \cr
Version: \tab 0.99.0 \cr
Date: \tab 2017-05-30 \cr
License: \tab Artistic-2.0 \cr
}
}
\examples{
genes <- list(K.demethylases = c("KDM1A", "KDM1B", "KDM2A"),
K.acetyltransferases = c("CLOCK", "CREBBP", "ELP3", "EP300"))

cancername <- "Breast Invasive Carcinoma (TCGA, Cell 2015)"

process.data.for.a.cancer(genes, cancername, "RNA-seq")

process.data.for.a.cancer(genes, cancername, "RNA-seq", desired.case.list = c(3,4,5),
data.presented.as = c("Frequency.Percentage", "Frequency.Ratio", "Mean.Expression"),
resolution=300, RowCex=1, ColCex=1, heatmapMargines=c(15,5),
cutoff=1.5, angle.for.heatmap.cancernames=30, heatmap.color = "redgreen")

}
\author{
Arman Shahrisa, \email{shahrisa.arman@hotmail.com} [maintainer, copyright holder]

Maryam Tahmasebi Birgani, \email{tahmasebi-ma@ajums.ac.ir}
}
