% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.multiple.studies.R
\name{process.multiple.studies}
\alias{process.multiple.studies}
\title{Checking Expression/methylation Profile for various cancer studies.}
\usage{
process.multiple.studies(genes, cancers, high.throughput.data.type,
data.presented.as = c("Frequency.Percentage", "Frequency.Ratio", "Mean.Value", "Median"),
shorteded.cancer.names = TRUE, genelimit="none", resolution=600, RowCex=0.8, ColCex=0.8,
heatmapMargines=c(24,17), cutoff=2, angle.for.heatmap.cancernames=45, heatmap.color = "RdBu",
reverse.heatmap.color = TRUE, rewrite.output.list = TRUE, round=TRUE, top.genes = TRUE,
validate.genes = TRUE, Use.CancerCode.as.Name = FALSE, simplify.visulization=FALSE,
simplifiction.cuttoff=FALSE)
}
\arguments{
\item{genes}{a list that contains at least one gene group}

\item{cancers}{a character vector or a matrix that containes desired cancer names. The character vector containes standard
cancer names that can be found on cbioportal.org, such as \code{Acute Myeloid Leukemia (TCGA, NEJM 2013)}. Alternatively,
a matrix can be used if users prefer user-defined cancer names, in which the first column of matrix comprises the standard
cancer names while the second column must contain the desired cancer names.}

\item{high.throughput.data.type}{a character string that is one of the following techniques: 'RNA-seq', 'microRNA-Seq',
'microarray.mRNA', 'microarray.microRNA' or 'methylation'.}

\item{data.presented.as}{a character vector that containes the statistical precedures users prefer the function to compute.
default is \code{c("Frequency.Percentage", "Frequency.Ratio", "Mean.Value", "Median")}. This will tell the function to
compute the following:
frequency precentage, which is the number of samples having value greather than specific cutoff
divided by the total sample size for each cancer study;
frequency ratio, which shows the number of selected samples divided by the total number of samples that give the frequency
percentage for each cancer-to know selecected and total sample sizes only;
Mean Expression, that contains mean value of selected samples for each cancer;
Median, which shows the median value of selected samples for each cancer.}

\item{shorteded.cancer.names}{a logical vector. If the value is set as true, function will try to remove the end part of
cancer names aiming to shorten them. The removed segment usually contains the name of scientific group that has conducted
the experiment.}

\item{genelimit}{if large number of genes exists in at least one gene group, this option can be use to limit the number of
genes to be shown on hitmap. For instance, \code{genelimit=50} will limit the heatmap to 50 genes showing the most variation.
The default value is \code{none}.}

\item{resolution}{a number. This option can be used to adjust the resolution of the output heatmaps as 'dot per inch'.
The defalut value is 600.}

\item{RowCex}{a number that specifies letter size in heatmap row names.}

\item{ColCex}{a number that specifies letter size in heatmap column names.}

\item{heatmapMargines}{a numeric vectors that can be use to set heatmap margins. The default value is
\code{heatmapMargines=c(15,07)}.}

\item{cutoff}{a number used to limit samples to those that are greather than specific number (cutoff). The default value for
methylation data is 0.6 while gene expression studies use default value of 2. For methylation studies, it is
\code{observed/expected ratio}, for the rest, it is \code{z-score} - Function can understand the correct unit. TO change the
cutoff to any desired number, change the option to \code{cutoff = desiredNumber} in which \code{desiredNumber} is the number
of interest.}

\item{angle.for.heatmap.cancernames}{a number that determines the angle with which the cancer names are shown in heatmaps.
The default value is 45 degree.}

\item{heatmap.color}{a character string matches standard color names. The default value is "RdBu". "redgreen" is also a popular
color in genomic studies. To see the rest of colors, please type \code{display.brewer.all()}.}

\item{reverse.heatmap.color}{a logical value that reverses the color gradiant for heatmap.}

\item{rewrite.output.list}{a logical value. This option can be used to modify heatmap, for instance change margin,
without obtaining data from internet again. If set to false, function will use the values previously stored in the
global environment to draw heatmaps and save excel file. The default value is \code{TRUE}.}

\item{round}{a logical value that, if set to be \code{TRUE}, will force the function to round all the calculated values
to two decimal places. The default value is \code{TRUE}.}

\item{top.genes}{a logical value that, if set as \code{TRUE}, cause the function to create three data.frame that contain the
gene names with the highest values for each cancer. To get all the three data.frames, `Frequency.Percentage`, `Mean.Value` and
`Median` must have been included for \code{data.presented.as}.}

\item{validate.genes}{a logical value that, if set to be \code{TRUE}, function will checks each cancer study to finds whether
or not each gene has a record. If a cancer study doesn't have a record for specific gene, it checks for alternative gene names
that cbioportal might use instead of the given gene name.}

\item{Use.CancerCode.as.Name}{a logical value that tells the function to use sandard abbreviated cancer names instead of complete cancer
names, if set to be \code{TRUE}. For example, \code{laml_tcga_pub} is the shortened name for \code{Acute Myeloid Leukemia (TCGA, NEJM 2013)}.}

\item{simplify.visulization}{a logical value that tells the function whether or not to change values under
\code{simplifiction.cuttoff} to zero. It only affects heatmaps to assist finding the candidate genes faster. Therefore, it is
not suited for publications.}

\item{simplifiction.cuttoff}{a logical value that, if \code{simplify.visulization = TRUE}, needs to be set as a desired cuttoff
for \code{simplify.visulization}. It has the same unit as \code{cutoff}.}
}
\value{
a list that containes some or all of the following groups, based on what user has chosen: \code{Validation.Results},
\code{Frequency.Percentage}, \code{Top.Genes.of.Frequency.Percentage}, \code{Frequency.Ratio}, \code{Mean.Value},
\code{Top.Genes.of.Mean.Value}, \code{Median}, \code{Top.Genes.of.Median}. It also saves these groups in one excel
file for convenience. Based on preference, three heatmaps for \code{Frequency.Percentage}, \code{Mean.Value} and
\code{Median} can be generated. If more than one group of genes is entered, output for each group will be strored in a separate sub-directory.
}
\description{
This function Obtains the requested data for the given genes across multiple cancer studies. It can check
whether or not all genes are included in cancer studies and, if not, looks for the alternative gene names. Tha main part
of function calculates frequency percentage, frequency ratio, mean expression and median of samples greather than specific
value in the selected cancer studies. Furthermore, it looks for the genes that comprise the highest values in each cancer study.
}
\details{
\tabular{lllll}{
Package: \tab cBioAutomatedTools \cr
Type: \tab Package \cr
Version: \tab 0.99.0 \cr
Date: \tab 2017-06-01 \cr
License: \tab Artistic-2.0 \cr
}
}
\examples{
# Creating a list that contains one gene group: 'K.acetyltransferases'
genes <- list(K.acetyltransferases = c("CLOCK", "CREBBP", "ELP3", "EP300"))

# creating a character vector of study names. alternatively, To see how to subset cancer names, please type '?available.cancer.types'
cancernames <- c("Acute Myeloid Leukemia (TCGA, Provisional)", "Adrenocortical Carcinoma (TCGA, Provisional)",
"Bladder Urothelial Carcinoma (TCGA, Provisional)", "Brain Lower Grade Glioma (TCGA, Provisional)",
"Breast Invasive Carcinoma (TCGA, Provisional)")

# Running the function to obtain and process the selected data
process.multiple.studies(genes, cancernames, "RNA-seq",
data.presented.as = c("Frequency.Percentage", "Frequency.Ratio", "Mean.Value"),
shorteded.cancer.names = FALSE, resolution=300, RowCex=1, ColCex=1, heatmapMargines=c(15,5),
cutoff=1.5, angle.for.heatmap.cancernames=30, heatmap.color = "redgreen")

}
\author{
Arman Shahrisa, \email{shahrisa.arman@hotmail.com} [maintainer, copyright holder]

Maryam Tahmasebi Birgani, \email{tahmasebi-ma@ajums.ac.ir}
}
