% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.one.study.R
\name{process.one.study}
\alias{process.one.study}
\title{Checking Expression/methylation Profile for various subgroups of a cancer study.}
\usage{
process.one.study(genes, cancername, high.throughput.data.type,
data.presented.as = c("Frequency.Percentage", "Frequency.Ratio", "Mean.Value", "Median"),
transposedHeatmap=FALSE, desired.case.list="None", genelimit="none",
resolution=600, RowCex=0.8, ColCex=0.8, heatmapMargines=c(10,10), cutoff="default",
angle.for.heatmap.cancernames=45, heatmap.color = "RdBu", reverse.heatmap.color = TRUE,
rewrite.output.list = TRUE, round=TRUE, top.genes = TRUE, validate.genes = TRUE,
Use.CancerCode.as.Name = FALSE, simplify.visulization=FALSE, simplifiction.cuttoff=FALSE)
}
\arguments{
\item{genes}{a list that contains at least one gene group}

\item{cancername}{a character string showing the desired cancer name. It is an standard cancer name that can be found on
cbioportal.org, such as \code{Acute Myeloid Leukemia (TCGA, NEJM 2013)}.}

\item{high.throughput.data.type}{a character string that is one of the following techniques: 'RNA-seq', 'microRNA-Seq',
'microarray.mRNA', 'microarray.microRNA' or 'methylation'.}

\item{data.presented.as}{a character vector that containes the statistical precedures users prefer the function to compute.
default is \code{c("Frequency.Percentage", "Frequency.Ratio", "Mean.Value", "Median")}. This will tell the function to
compute the following:
frequency precentage, which is the number of samples having value greather than specific cutoff divided by the total sample
size for each cancer;
frequency ratio, which shows the number of selected samples divided by the total number of samples that give the frequency
percentage
for each cancer-to know selecected and total sample sizes only;
Mean Expression, that contains mean value of selected samples for each cancer;
Median, which shows the median value of selected samples for each cancer.}

\item{transposedHeatmap}{a logical value that can exchange the rows and columns of heatmaps.}

\item{desired.case.list}{a numeric vector that contains the index of rdesired cancer subgroups. If set to be \code{FALSE},
function will ask the user to enter them during the process. The default value is \code{FALSE}.}

\item{genelimit}{if large number of genes exists in at least one gene group, this option can be use to limit the number of
genes to be shown on hitmap. For instance, \code{genelimit=50} will limit the heatmap to 50 genes showing the most variation.
The default value is \code{FALSE}.}

\item{resolution}{a number. This option can be used to adjust the resolution of the output heatmaps as 'dot per inch'. The
defalut value is 600.}

\item{RowCex}{a number that specifies letter size in heatmap row names.}

\item{ColCex}{a number that specifies letter size in heatmap column names.}

\item{heatmapMargines}{a numeric vectors that can be use to set heatmap margins. The default value is
\code{heatmapMargines=c(10,10)}.}

\item{cutoff}{a number used to limit samples to those that are greather than specific number (cutoff). The default value for
methylation data is 0.6 while gene expression studies use default value of 2. For methylation studies, it is
\code{observed/expected ratio}, for the rest, it is \code{z-score} - Function can understand the correct unit. TO change the
cutoff to any desired number, change the option to \code{cutoff = desiredNumber} in which \code{desiredNumber} is the number
of interest.}

\item{angle.for.heatmap.cancernames}{a number that determines the angle with which the cancer names are shown in heatmaps.
The default value is 45 degree.}

\item{heatmap.color}{a character string matches standard color names. The default value is "RdBu". "redgreen" is also a popular
color in genomic studies. To see the rest of colors, please type \code{display.brewer.all()}.}

\item{reverse.heatmap.color}{a logical value that reverses the color gradiant for heatmap.}

\item{rewrite.output.list}{a logical value. This option can be used to modify heatmap, for instance change margin,
without obtaining data from internet again. If set to false, function will use the values previously stored in the
global environment to draw heatmaps and save excel file. The default value is \code{TRUE}.}

\item{round}{a logical value that, if set to be \code{TRUE}, will force the function to round all the calculated values
to two decimal places. The default value is \code{TRUE}.}

\item{top.genes}{a logical value that, if set as \code{TRUE}, cause the function to create three data.frame that contain the
gene names with the highest values for each cancer. To get all the three data.frames, `Frequency.Percentage`, `Mean.Value` and
`Median` must have been included for \code{data.presented.as}.}

\item{validate.genes}{a logical value that, if set to be \code{TRUE}, function will checks each cancer study to finds whether
or not each gene has a record. If the given cancer doesn't have a record for specific gene, it checks for alternative gene
names that cbioportal might use instead of the given gene name.}

\item{Use.CancerCode.as.Name}{a logical value that tells the function to use abbreviated cancer names instead of complete
cancer names, if set to be \code{TRUE}. For example, \code{laml_tcga_pub} is the shortened name for
\code{Acute Myeloid Leukemia (TCGA, NEJM 2013)}.}

\item{simplify.visulization}{a logical value that tells the function whether or not to change values under
\code{simplifiction.cuttoff} to zero. It only affects heatmaps to assist finding the candidate genes faster. Therefore, it is
not suited for publications.}

\item{simplifiction.cuttoff}{a logical value that, if \code{simplify.visulization = TRUE}, needs to be set as a desired cuttoff
for \code{simplify.visulization}. It has the same unit as \code{cutoff}.}
}
\value{
a list that containes some or all of the following groups, based on what user has chosen: \code{Validation.Results},
\code{Frequency.Percentage}, \code{Top.Genes.of.Frequency.Percentage}, \code{Frequency.Ratio}, \code{Mean.Value},
\code{Top.Genes.of.Mean.Value}, \code{Median}, \code{Top.Genes.of.Median}. It also saves these groups in one excel
file for convenience. Based on preference, three heatmaps for \code{Frequency.Percentage}, \code{Mean.Value} and
\code{Median} can be generated. If more than one gene group is entered, output for each group will be strored in a separate sub-directory.
}
\description{
This function Obtaines the requested data for the given genes across multiple subgroups of a cancer. It can
check whether or not all genes are included in subgroups of a cancer study and, if not, looks for the alternative gene names.
Tha main part of function calculates frequency percentage, frequency ratio, mean expression and median of samples greather than
specific value in the selected subgroups of the cancer. Furthermore, it looks for the genes that comprise the highest values
in each cancer subgroup.
}
\details{
\tabular{lllll}{
Package: \tab cBioAutomatedTools \cr
Type: \tab Package \cr
Version: \tab 0.99.0 \cr
Date: \tab 2017-05-31 \cr
License: \tab Artistic-2.0 \cr
}
}
\examples{
genes <- list(K.demethylases = c("KDM1A", "KDM1B", "KDM2A"),
K.acetyltransferases = c("CLOCK", "CREBBP", "ELP3", "EP300"))

cancername <- "Breast Invasive Carcinoma (TCGA, Cell 2015)"

process.one.study(genes, cancername, "RNA-seq")

process.one.study(genes, cancername, "RNA-seq", desired.case.list = c(3,4,5),
data.presented.as = c("Frequency.Percentage", "Frequency.Ratio", "Mean.Value"),
resolution=300, RowCex=1, ColCex=1, heatmapMargines=c(15,5),
cutoff=1.5, angle.for.heatmap.cancernames=30, heatmap.color = "redgreen")

}
\author{
Arman Shahrisa, \email{shahrisa.arman@hotmail.com} [maintainer, copyright holder]

Maryam Tahmasebi Birgani, \email{tahmasebi-ma@ajums.ac.ir}
}
