% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbaf-automatedStatistics.R
\name{automatedStatistics}
\alias{automatedStatistics}
\title{Performing the requested statistics for various studies / subgroups of a study.}
\usage{
automatedStatistics(submissionName, obtainedDataType = "multiple studies",
calculate = c("frequencyPercentage", "frequencyRatio", "meanValue", "medianValue"),
cutoff=NULL, round=TRUE, topGenes = TRUE)
}
\arguments{
\item{submissionName}{a character string containing name of interest. It is used for naming the process.}

\item{obtainedDataType}{a character string that specifies the type of input data to function. Two options are availabe:
'single study' and 'multiple studies'. The function uses 'obtainedDataType' and 'submissionName' to construct
the name of input data. Default value is 'multiple studies'.}

\item{calculate}{a character vector that containes the statistical precedures users prefer the function to compute.
Default input is \code{c("frequencyPercentage", "frequencyRatio", "Mean.Value", "medianValue")}. This will tell the function to
compute the following:
'frequencyPercentage', which is the number of samples having the value greather than specific cutoff divided by the total sample
size for every study / study subgroup;
'frequency ratio', which shows the number of selected samples divided by the total number of samples that give the frequency
percentage for every study / study subgroup -to know selecected and total sample sizes only;
'Mean Value', that contains mean value of selected samples for each study;
'Median Value', which shows the median value of selected samples for each study.}

\item{cutoff}{a number used to limit samples to those that are greather than specific number (cutoff). The default value for
methylation data is 0.6 while gene expression studies use default value of 2. For methylation studies, it is
\code{observed/expected ratio}, for the rest, it is 'z-score'. TO change the cutoff to any desired number, change the
option to \code{cutoff = desiredNumber} in which desiredNumber is the number of interest.}

\item{round}{a logical value that, if set to be TRUE, will force the function to round all the calculated values
to two decimal places. The default value is TRUE.}

\item{topGenes}{a logical value that, if set as TRUE, causes the function to create three data.frame that contain the five
top genes for each cancer. To get all the three data.frames, "Frequency.Percentage", "Mean.Value" and "Median" must have been
included for \code{calculate}.}
}
\value{
a list that contains some or all of the following statistical measurements for every gene group, based on what user has chosen:
\code{Frequency.Percentage}, \code{Top.Genes.of.Frequency.Percentage}, \code{Frequency.Ratio}, \code{Mean.Value}, \code{Top.Genes.of.Mean.Value},
\code{Median}, \code{Top.Genes.of.Median}.
}
\description{
This function calculates frequency percentage, frequency ratio, mean value and median value of samples greather than
specific cutoff in the selected study / subgroups of the study. Furthermore, it can looks for the five genes that contain the highest values
in each study / study subgroup. It uses the data generated by obtainOneStudy()/obtainMultipleStudies() function.
}
\details{
\tabular{lllll}{
Package: \tab cbaf \cr
Type: \tab Package \cr
Version: \tab 0.99.0 \cr
Date: \tab 2017-07-30 \cr
License: \tab Artistic-2.0 \cr
}
}
\examples{
genes <- list(K.demethylases = c("KDM1A", "KDM1B", "KDM2A"))

obtainOneStudy(genes, "test", "Breast Invasive Carcinoma (TCGA, Cell 2015)",
"RNA-seq", desiredCaseList = c(3,4))

automatedStatistics("test", obtainedDataType = "single study", calculate =
c("frequencyPercentage", "frequencyRatio"))

}
\author{
Arman Shahrisa, \email{shahrisa.arman@hotmail.com} [maintainer, copyright holder]

Maryam Tahmasebi Birgani, \email{tahmasebi-ma@ajums.ac.ir}
}
