% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.data.types.R
\name{available.data.types}
\alias{available.data.types}
\title{Checking which Datasets are available for Each Cancer study.}
\usage{
available.data.types()
}
\value{
A matrix that contain all cancers versus available data types. It is stored in the
global enviroment (user's workspace) as 'available.data.types.output'. For convenience, an
excel file - 'Available Data Types Output.xlsx' - is also generated in the working directory.
}
\description{
This function checks all the cancer studies that are registered in 'cbioportal.org' to
examine whether or not they contain RNA-seq, microRNA-seq, microarray(mRNA),
microarray(miRNA) and methylation data.
}
\details{
\tabular{lllll}{
Package: \tab cBioAutomatedTools \cr
Type: \tab Package \cr
Version: \tab 0.99.0 \cr
Date: \tab 2017-06-01 \cr
License: \tab Artistic-2.0 \cr
}
}
\examples{
# After viewing 'available.data.types.output', users can create a character vector of desired study names
# for process.multiple.studies() or pick one study for process.one.study(). For mor information,
# please type '?process.multiple.studies' and '?process.one.study'.
cancernames <- c("Acute Myeloid Leukemia (TCGA, Provisional)", "Adrenocortical Carcinoma (TCGA, Provisional)",
"Bladder Urothelial Carcinoma (TCGA, Provisional)", "Brain Lower Grade Glioma (TCGA, Provisional)",
"Breast Invasive Carcinoma (TCGA, Provisional)")

}
\author{
Arman Shahrisa, \email{shahrisa.arman@hotmail.com} [maintainer, copyright holder]

Maryam Tahmasebi Birgani, \email{tahmasebi-ma@ajums.ac.ir}
}
