% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbaf-availableData.R
\name{availableData}
\alias{availableData}
\title{Checking which Data types are available for Each Cancer study.}
\usage{
availableData(outputName, excelFile = TRUE)
}
\arguments{
\item{outputName}{a character string that is required to name the output and, if requested, excel file.}

\item{excelFile}{a logical value that tells the function whether or not export the results as an excel file. Default vlue is TRUE}
}
\value{
A matrix that contain all cancer studies versus available data types. It is stored in the
global enviroment (user's workspace) as a variable according to input name. For convenience, an
excel file is also generated with the same name in the working directory.
}
\description{
This function checks all the cancer studies that are registered in 'cbioportal.org' to
examine whether or not they contain RNA-seq, microRNA-seq, microarray(mRNA), microarray(miRNA)
and methylation data.
}
\details{
\tabular{lllll}{
Package: \tab cbaf \cr
Type: \tab Package \cr
Version: \tab 0.99.0 \cr
Date: \tab 2017-06-28 \cr
License: \tab Artistic-2.0 \cr
}
}
\examples{
availableData(outputName = "list.of.studies", excelFile = FALSE)

}
\author{
Arman Shahrisa, \email{shahrisa.arman@hotmail.com} [maintainer, copyright holder]

Maryam Tahmasebi Birgani, \email{tahmasebi-ma@ajums.ac.ir}
}
