% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbaf-availableData.R
\name{availableData}
\alias{availableData}
\title{Check which Data types are available for each cancer study.}
\usage{
availableData(excelFileName)
}
\arguments{
\item{excelFileName}{a character string that is required to name the output
and, if requested, excel file.}
}
\value{
A matrix that contains all cancer studies versus available data types
. It is available as an excel file with the given name.
}
\description{
This function checks all the cancer studies that are registered
in 'cbioportal.org' to examine whether or not they contain RNA-seq,
microRNA-seq, microarray(mRNA), microarray(miRNA) and methylation data.
}
\details{
\tabular{lllll}{
Package: \tab cbaf \cr
Type: \tab Package \cr
Version: \tab 1.0.0 \cr
Date: \tab 2017-09-10 \cr
License: \tab Artistic-2.0 \cr
}
}
\author{
Arman Shahrisa, \email{shahrisa.arman@hotmail.com} [maintainer, copyright holder]

Maryam Tahmasebi Birgani, \email{tahmasebi-ma@ajums.ac.ir}
}
