% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbaf-availableData.R
\name{availableData}
\alias{availableData}
\title{Check which Data types are available for each cancer study.}
\usage{
availableData(excelFileName)
}
\arguments{
\item{excelFileName}{a character string that is required to name the output
and, if requested, excel file.}
}
\value{
An excel file that contains all the cancer studies versus available
data types
}
\description{
This function checks all the cancer studies that are registered
in 'cbioportal.org' to examine whether or not they contain RNA-Seq,
microRNA-Seq, microarray(mRNA), microarray(miRNA) and methylation data.
}
\details{
\tabular{lllll}{
Package: \tab cbaf \cr
Type: \tab Package \cr
Version: \tab 1.13.2 \cr
Date: \tab 2020-12-22 \cr
License: \tab Artistic-2.0 \cr
}
}
\author{
Arman Shahrisa, \email{shahrisa.arman@hotmail.com} [maintainer,
copyright holder]

Maryam Tahmasebi Birgani, \email{tahmasebi-ma@ajums.ac.ir}
}
