% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbaf-cleanDatabase.R
\name{cleanDatabase}
\alias{cleanDatabase}
\title{Clean the created database(s)}
\usage{
cleanDatabase(databaseNames = NULL)
}
\arguments{
\item{databaseNames}{a character vector that contains name of databases that
will be removed. The default value in \code{null}.}
}
\value{
prints the number of removed databases.
}
\description{
This function removes the created databases in the cbaf package
directory. This helps users to obtain the fresh data from cbioportal.org.
}
\details{
\tabular{lllll}{
Package: \tab cbaf \cr
Type: \tab Package \cr
Version: \tab 1.9.1 \cr
Date: \tab 2020-01-01 \cr
License: \tab Artistic-2.0 \cr
}
}
\examples{
cleanDatabase(databaseNames = "Whole")



}
\author{
Arman Shahrisa, \email{shahrisa.arman@hotmail.com} [maintainer,
copyright holder]

Maryam Tahmasebi Birgani, \email{tahmasebi-ma@ajums.ac.ir}
}
