% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbaf-heatmapOutput.R
\name{heatmapOutput}
\alias{heatmapOutput}
\title{Generating heatmaps for various studies/subgroups of a study.}
\usage{
heatmapOutput(submissionName, shortenStudyNames = TRUE, genelimit = "none", resolution = 600, RowCex = 0.8,
ColCex = 0.8, heatmapMargines = c(10,10), angleForYaxisNames = 45, heatmapColor = "RdBu", reverseColor = TRUE,
transposedHeatmap = FALSE, simplify = FALSE, simplifictionCuttoff = FALSE)
}
\arguments{
\item{submissionName}{a character string containing name of interest. It is used for naming the process.}

\item{shortenStudyNames}{a logical vector. If the value is set as true, function will try to remove the end part of
cancer names aiming to shorten them. The removed segment usually contains the name of scientific group that has conducted
the experiment.}

\item{genelimit}{if large number of genes exist in at least one gene group, this option can be used to limit the number of
genes that are shown on hitmap. For instance, \code{genelimit=50} will limit the heatmap to 50 genes showing the most variation
across multiple study / study subgroups.
The default value is \code{none}.}

\item{resolution}{a number. This option can be used to adjust the resolution of the output heatmaps as 'dot per inch'.
The defalut value is 600.}

\item{RowCex}{a number that specifies letter size in heatmap row names.}

\item{ColCex}{a number that specifies letter size in heatmap column names.}

\item{heatmapMargines}{a numeric vector that can be used to set heatmap margins. The default value is
\code{heatmapMargines=c(15,07)}.}

\item{angleForYaxisNames}{a number that determines the angle with which the studies/study subgroups names are shown in heatmaps.
The default value is 45 degree.}

\item{heatmapColor}{a character string that defines heatmao color. The default value is "RdBu". "redgreen" is also a popular
color in genomic studies. To see the rest of colors, please type \code{display.brewer.all()}.}

\item{reverseColor}{a logical value that reverses the color gradiant for heatmap(s).}

\item{transposedHeatmap}{a logical value that changes row and colums of heatmap.}

\item{simplify}{a logical value that tells the function whether or not to change values under
\code{simplifiction.cuttoff} to zero. The purpose behind this option is to facilitate seeing candidate genes. Therefore, it is
not suited for publications.}

\item{simplifictionCuttoff}{a logical value that, if \code{simplify.visulization = TRUE}, needs to be set as a desired cuttoff
for \code{simplify.visulization}. It has the same unit as \code{cutoff}.}
}
\value{
Based on preference, three heatmaps for "Frequency.Percentage", "Mean.Value" and "Median.value" can be
generated. If more than one group of genes are entered, output for each group will be strored in a separate sub-directory.
}
\description{
This function can prepare heatmap for 'frequency percentage', 'mean value' and 'median value' data provided by
automatedStatistics() function.
}
\details{
\tabular{lllll}{
Package: \tab cbaf \cr
Type: \tab Package \cr
Version: \tab 0.99.0 \cr
Date: \tab 2017-06-28 \cr
License: \tab Artistic-2.0 \cr
}
}
\examples{
heatmapOutput(submissionName = "test")

}
\author{
Arman Shahrisa, \email{shahrisa.arman@hotmail.com} [maintainer, copyright holder]

Maryam Tahmasebi Birgani, \email{tahmasebi-ma@ajums.ac.ir}s
}
