% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbaf-obtainMultipleStudies.R
\name{obtainMultipleStudies}
\alias{obtainMultipleStudies}
\title{Obtaining the requested data for various cancer studies.}
\usage{
obtainMultipleStudies(genesList, submissionName, studiesNames, desiredTechnique,
cancerCode = FALSE, validateGenes = TRUE)
}
\arguments{
\item{genesList}{a list that contains at least one gene group}

\item{submissionName}{a character string containing name of interest. It is used for naming the process.}

\item{studiesNames}{a character vector or a matrix that containes desired cancer names. The character vector containes standard
names of cancer studies that can be found on cbioportal.org, such as 'Acute Myeloid Leukemia (TCGA, NEJM 2013)'. Alternatively,
a matrix can be used if users prefer user-defined cancer names, in which the first column of matrix comprises the standard
cancer names while the second column must contain the desired cancer names.}

\item{desiredTechnique}{a character string that is one of the following techniques: 'RNA-seq', 'microRNA-Seq', 'microarray.mRNA',
'microarray.microRNA' or 'methylation'.}

\item{cancerCode}{a logical value that tells the function to use cbioportal abbreviated cancer names instead of complete cancer
names, if set to be 'TRUE'. For example, 'laml_tcga_pub' is the abbreviated name for 'Acute Myeloid Leukemia (TCGA, NEJM 2013)'.}

\item{validateGenes}{a logical value that, if set to be \code{TRUE}, function will check each cancer study to find whether
or not each gene has a record. If a cancer study doesn't have a record for specific gene, it checks for alternative gene names
that cbioportal might use instead of the given gene name.}
}
\value{
a list that contains the obtained data without further processing. Name of the list starts with 'obS' and contains
submissionName. Inside the list, there is one subgroup for every gene group, which itself contains one matrix for every cancer study.
In addition, if validateGenes = TRUE, a secondary list containing gene validation results will be stored. Name of the second
list starts with 'vaS' and contains submissionName.
}
\description{
This function Obtains the requested data for the given genes across multiple cancer studies. It can check
whether or not all genes are included in cancer studies and, if not, looks for the alternative gene names.
}
\details{
\tabular{lllll}{
Package: \tab cbaf \cr
Type: \tab Package \cr
Version: \tab 0.99.0 \cr
Date: \tab 2017-06-28 \cr
License: \tab Artistic-2.0 \cr
}
}
\examples{
genes <- list(K.acetyltransferases = c("CLOCK", "CREBBP", "ELP3", "EP300"))

studies <- c("Acute Myeloid Leukemia (TCGA, Provisional)", "Adrenocortical Carcinoma (TCGA, Provisional)")

# obtainMultipleStudies(genes, "test2", studies, "RNA-seq")

}
\author{
Arman Shahrisa, \email{shahrisa.arman@hotmail.com} [maintainer, copyright holder]

Maryam Tahmasebi Birgani, \email{tahmasebi-ma@ajums.ac.ir}
}
