% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbaf-processOneStudy.R
\name{processOneStudy}
\alias{processOneStudy}
\title{Check Expression/methylation Profile for various subgroups of a cancer
 study.}
\usage{
processOneStudy(genesList, submissionName, studyName, desiredTechnique
  , desiredCaseList = FALSE, validateGenes = TRUE, calculate =
  c("frequencyPercentage", "frequencyRatio", "meanValue"), cutoff=NULL,
  round=TRUE, topGenes = TRUE, shortenStudyNames = TRUE, genelimit = "none",
  resolution = 600, RowCex = 0.8, ColCex = 0.8, heatmapMargines = c(15,07),
  angleForYaxisNames = 45, heatmapColor = "RdBu", reverseColor = TRUE,
  transposedHeatmap = FALSE, simplify = FALSE, simplifictionCuttoff = FALSE,
  genesToDrop = NULL)
}
\arguments{
\item{genesList}{a list that contains at least one gene group}

\item{submissionName}{a character string containing name of interest. It is
used for naming the process.}

\item{studyName}{a character string showing the desired cancer name. It is an
standard cancer study name that can be found on cbioportal.org, such as
'Acute Myeloid Leukemia (TCGA, NEJM 2013)'.}

\item{desiredTechnique}{a character string that is one of the following
techniques: 'RNA-seq', 'microRNA-Seq', 'microarray.mRNA',
'microarray.microRNA' or 'methylation'.}

\item{desiredCaseList}{a numeric vector that contains the index of desired
cancer subgroups, assuming the user knows index of desired subgroups. If not,
 desiredCaseList has been set as 'none', function will show the available
 subgroups and ask the user to enter the desired ones during the process. The
  default value is 'none'.}

\item{validateGenes}{a logical value that, if set to be 'TRUE', causes the
function to check each cancer study to find whether or not each gene has a
record. If a cancer doesn't have a record for specific gene, function looks
for alternative gene names that cbioportal might use instead of the given
gene name.}

\item{calculate}{a character vector that containes the statistical procedures
 users prefer the function to compute. The complete input is
\code{c("frequencyPercentage", "frequencyRatio", "meanValue", "medianValue")}
 . This will tell the function to compute the following:
'frequencyPercentage', which is the percentge of samples having the value
greather than specific cutoff divided by the total sample size for every
study / study subgroup;
'frequency ratio', which shows the number of selected samples divided by the
total number of samples that give the frequency percentage for every study /
study subgroup. It shows the selected and total sample sizes.;
'Mean Value', that contains mean value of selected samples for each study;
'Median Value', which shows the median value of selected samples for each
study.
The default input is \code{calculate = c("frequencyPercentage",
"frequencyRatio", "meanValue")}.}

\item{cutoff}{a number used to limit samples to those that are greather than
specific number (cutoff). The default value for methylation data is 0.6 while
 gene expression studies use default value of 2. For methylation studies, it
 is \code{observed/expected ratio}, for the rest, it is 'z-score'. To change
 the cutoff to any desired number, change the option to
 \code{cutoff = desiredNumber}, in which desiredNumber is the number of
 interest.}

\item{round}{a logical value that, if set to be TRUE, will force the function
to round all the calculated values to two decimal places. The default value
is TRUE.}

\item{topGenes}{a logical value that, if set as TRUE, causes the function to
create three data.frame that contain the five top genes for each cancer. To
get all the three data.frames, "frequencyPercentage", "meanValue" and
"medianValue" must have been included for \code{calculate}.}

\item{shortenStudyNames}{a logical vector. If the value is set as TRUE,
function will try to remove the last part of the cancer names aiming to
shorten them. The removed segment usually contains the name of scientific
group that has conducted the experiment.}

\item{genelimit}{if large number of genes exist in at least one gene group,
this option can be used to limit the number of genes that are shown on
heatmap. For instance, \code{genelimit=50} will limit the heatmap to 50 genes
 showing the most variation across multiple study / study subgroups. The
 default value is \code{none}.}

\item{resolution}{a number. This option can be used to adjust the resolution
of the output heatmaps as 'dot per inch'. The defalut value is 600.}

\item{RowCex}{a number that specifies letter size in heatmap row names.}

\item{ColCex}{a number that specifies letter size in heatmap column names.}

\item{heatmapMargines}{a numeric vector that is used to set heatmap margins.
The default value is
\code{heatmapMargines=c(15,07)}.}

\item{angleForYaxisNames}{a number that determines the angle with which the
studies/study subgroups names are shown in heatmaps. The default value is 45
degree.}

\item{heatmapColor}{a character string that defines heatmap color. The
default value is "RdBu". "redgreen" is also a popular color in genomic
studies. To see the rest of colors, please type \code{library(RColorBrewer)}
and then \code{display.brewer.all()}.}

\item{reverseColor}{a logical value that reverses the color gradiant for
heatmap(s).}

\item{transposedHeatmap}{a logical value that transposes heatmap rows to
columns and vice versa.}

\item{simplify}{a logical value that tells the function whether or not to
change values under \code{simplifiction.cuttoff} to zero. The purpose behind
this option is to facilitate seeing candidate genes. Therefore, it is not
suited for publications. Default value is 'FALSE'.}

\item{simplifictionCuttoff}{a logical value that, if
\code{simplify.visulization = TRUE}, needs to be set as a desired cuttoff for
 \code{simplify.visulization}. It has the same unit as \code{cutoff}.}

\item{genesToDrop}{a character vector. Gene names within this vector will be
omitted from heatmap.}
}
\value{
a BiocFileCache object that containes some or all of the following
groups, based on what user has chosen: ObtainedData, validationResults,
frequencyPercentage, Top.Genes.of.Frequency.Percentage, frequencyRatio,
meanValue, Top.Genes.of.Mean.Value, medianValue, Top.Genes.of.Median.Value.
It also saves these results in one excel files for convenience.
Based on preference, three heatmaps for frequency percentage, mean value and
median can be generated. If more than one group of genes is entered, output
for each group will be strored in a separate sub-directory.
}
\description{
This function Obtains the requested data for the given genes
across multiple subgroups of a cancer. It can check whether or not all genes
are included in subgroups of a cancer study and, if not, looks for the
alternative gene names. Then it calculates frequency percentage, frequency
ratio, mean value and median value of samples greather than specific value in
 the selected subgroups of the cancer. Furthermore, it looks for the five
 genes that comprise the highest values in each cancer study subgroup.
}
\details{
\tabular{lllll}{
Package: \tab cbaf \cr
Type: \tab Package \cr
Version: \tab 1.0.0 \cr
Date: \tab 2017-09-26 \cr
License: \tab Artistic-2.0 \cr
}
}
\examples{
genes <- list(K.demethylases = c("KDM1A", "KDM1B", "KDM2A", "KDM2B", "KDM3A",
 "KDM3B", "JMJD1C", "KDM4A"), K.methyltransferases = c("SUV39H1", "SUV39H2",
 "EHMT1", "EHMT2", "SETDB1", "SETDB2", "KMT2A", "KMT2A"))

processOneStudy(genes, "test", "Breast Invasive Carcinoma (TCGA, Cell 2015)",
"RNA-seq", desiredCaseList = c(2,3,4,5), calculate = c("frequencyPercentage",
"frequencyRatio"), heatmapMargines = c(16, 10), RowCex = 1, ColCex = 1)

}
\author{
Arman Shahrisa, \email{shahrisa.arman@hotmail.com} [maintainer,
copyright holder]

Maryam Tahmasebi Birgani, \email{tahmasebi-ma@ajums.ac.ir}
}
