% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbaf-xlsxOutput.R
\name{xlsxOutput}
\alias{xlsxOutput}
\title{Generating excel output for various studies/subgroups of a study.}
\usage{
xlsxOutput(submissionName)
}
\arguments{
\item{submissionName}{a character string containing name of interest. It is used for naming the process.}
}
\value{
It generates one excel file for each genegroup.
}
\description{
This function generates excel files containing gene validation and all selected statistical methods. It
uses outputs of obtainOneStudy()/obtainMultipleStudies() and automatedStatistics() functions.
}
\details{
\tabular{lllll}{
Package: \tab cBioAutomatedTools \cr
Type: \tab Package \cr
Version: \tab 0.99.0 \cr
Date: \tab 2017-06-22 \cr
License: \tab Artistic-2.0 \cr
}
}
\author{
Arman Shahrisa, \email{shahrisa.arman@hotmail.com} [maintainer, copyright holder]

Maryam Tahmasebi Birgani, \email{tahmasebi-ma@ajums.ac.ir}
}
